% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guttman.R
\name{getmode}
\alias{getmode}
\title{Return mode (dominant) value from numerical vector}
\usage{
getmode(v)
}
\arguments{
\item{v}{numeric vector}
}
\description{
Return mode (dominant) value from numerical vector
}

