% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guttman.R
\name{guttman}
\alias{guttman}
\title{Create guttman object}
\usage{
guttman(dataset)
}
\arguments{
\item{dataset}{a matrix or data frame with binary values}
}
\value{
Object of \emph{guttman} class, a list which contains:
\describe{
    \item{dataset}{data frame, original dataset provided as imput}
    \item{optimal_dataset}{data frame, dataset sorted in order to minimize Guttman errors}
    \item{optimal_row_errors}{numeric vector of row errors for optimal_dataset}
    \item{optimal_column_errors}{numeric vector of column errors for optimal_dataset}
    \item{total_errors}{numeric, total number of Guttman errors in optimal_dataset}
    \item{TN}{numeric, total number of mode categories in columns}
    \item{N}{numeric, total number of responses}
    \item{MMR}{numeric, Minimal Marginal Reproducibility}
    \item{ME}{numeric, Marginal Error}
    \item{CR}{numeric, Coefficient of Reproducibility}
    \item{CS}{numeric, Coefficient of Scalability }
}
}
\description{
Create guttman object
}
\examples{
data("SFd")
guttman_object = guttman(dataset = SFd)
summary(guttman_object)
}

